@@ -5,19 +5,16 @@
   "basename": "/",
   "dependencies": {
     "@babel/core": "7.4.3",
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
     "@reduxjs/toolkit": "^1.9.3",
     "@svgr/webpack": "4.1.0",
     "@typescript-eslint/eslint-plugin": "1.6.0",
     "@typescript-eslint/parser": "1.6.0",
-    "axios": "^1.3.4",
     "babel-eslint": "10.0.1",
     "babel-jest": "^24.9.0",
     "babel-loader": "8.0.5",
     "babel-plugin-named-asset-import": "^0.3.7",
     "babel-preset-react-app": "^9.1.2",
-    "bootstrap": "^4.6.1",
+    "bootstrap": "^4.6.2",
     "camelcase": "^5.2.0",
     "case-sensitive-paths-webpack-plugin": "2.2.0",
     "css-loader": "2.1.1",
@@ -36,7 +33,6 @@
     "file-loader": "3.0.1",
     "firebase": "^9.17.1",
     "formik": "^2.2.9",
-    "framer-motion": "^6.5.1",
     "fs-extra": "7.0.1",
     "google-map-react": "^1.1.7",
     "html-webpack-plugin": "4.0.0-beta.5",
@@ -54,14 +50,14 @@
     "postcss-normalize": "7.0.1",
     "postcss-preset-env": "6.6.0",
     "postcss-safe-parser": "4.0.1",
-    "react": "^17.0.2",
+    "react": "^18.0.0",
     "react-accessible-accordion": "^3.3.5",
     "react-animate-on-scroll": "^2.1.5",
     "react-app-polyfill": "^1.0.6",
-    "react-bootstrap": "^1.6.4",
+    "react-bootstrap": "^2.7.2",
     "react-countup": "^4.4.0",
     "react-dev-utils": "^9.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.0.0",
     "react-helmet": "^5.2.1",
     "react-icons": "^3.11.0",
     "react-image-lightbox": "^5.1.4",
